[MRG] Create media and offering pages (#89)
This PR adds media and services pages. It also rebuilds the entire layout for all pages and updates all pages to use HUGO shortcodes (instead of HUGO partials). By switching to shortcodes, many files could be removed from the repository. The result is a lightweight version that is more flexible and powerful than the previous one.