Reorganise privacy policies to allow archiving (#268)
This is largely based on https://policies.google.com/privacy. Which shows the current version, links to the archive and lets you link to particular past versions.

With this we have the following pages:

- current privacy policy with link to archive https://deploy-preview-268--skribble.netlify.com/privacy/
- past versions/archive https://deploy-preview-268--skribble.netlify.com/privacy/archive/
- each version has a stable link that will always be available/fixed (can be used in contracts):
  - https://deploy-preview-268--skribble.netlify.com/privacy/archive/20191022/
  - https://deploy-preview-268--skribble.netlify.com/privacy/archive/20181018/
- pages should continue to be available under their old slugs like https://deploy-preview-268--skribble.netlify.com/privacy-v2018-10-18, which redirect the user to the new address

When we add a new policy we have to create a privacy-20200401.md, fill it with content, update privacy-archive.md, update the YAML header of the privacy document that was current until now to move it to a stable slug and remove the alias, setup slug and alias for the new current terms.

If we like this setup and organisation I'd propose we move GTCs, ToUs and privacy policies in all three languages to this structure.