fix(Cookie consent reset): Fixes cookie consent resetting (#439)
The built in function from klaro (our cookie consent library) to reset cookies doesn't seem play well with explicitly defined domains.

This PR replaces that function with a "brute force" deletion of all cookies currently accessible to the browser window.

Does not mess with the only cookie which is really relevant to safeguard in our cookie flow. This would be x-access-token from Wave. It won't touch that one, since it has domain my.skribble.com and not .skribble.com. So, it's not accesible.

Unfortunately it can only be fully tested once it reaches production.